// client/src/__generated__/todos.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: todos
// ====================================================

export interface todos_todos {
  id: string;
  text: string;
  complete: boolean;
}

export interface todos {
  todos: todos_todos[] | null;
}

//==============================================================
// START Enums and Input Objects
// All enums and input objects are included in every output file
// for now, but this will be changed soon.
// TODO: Link to issue to fix this.
//==============================================================

//==============================================================
// END Enums and Input Objects
//==============================================================