// client/src/__generated__/updateComplete.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: updateComplete
// ====================================================

export interface updateComplete {
  updateComplete: boolean;
}

export interface updateCompleteVariables {
  id: string;
  complete: boolean;
}

//==============================================================
// START Enums and Input Objects
// All enums and input objects are included in every output file
// for now, but this will be changed soon.
// TODO: Link to issue to fix this.
//==============================================================

//==============================================================
// END Enums and Input Objects
//==============================================================